Added some keypresses so that now you can scroll the list up and down. Hitting escape will exit the program. Hitting page up or page down will also scroll by screen rather than by item.
Two errors: one, don't hit space. It tries to open the menu, and proceeds to not do anything about it. Two, the screen gets stuck at the very top if you scroll up all the way. You can unhook it by hitting page down. I'll look into this. It should be relative to the cursor later.